tweak project IDs, fix the typo in plugability
