+++ impl-tck/tck-tests-pluggability/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2020, 2021 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2020, 2022 Oracle and/or its affiliates. All rights reserved.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -24,16 +24,16 @@
 
     <parent>
         <groupId>org.eclipse.jsonp</groupId>
-        <artifactId>jakarta.json-tck</artifactId>
+        <artifactId>tck-project</artifactId>
         <version>2.1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>jakarta.json-tck-tests-plugability</artifactId>
+    <artifactId>jakarta.json-tck-tests-pluggability</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>jakarta.json</groupId>
-            <artifactId>jakarta.json-tck-tests-plugability</artifactId>
+            <artifactId>jakarta.json-tck-tests-pluggability</artifactId>
             <version>${jsonp-tck.version}</version>
             <scope>test</scope>
         </dependency>
@@ -52,7 +52,7 @@
                 <version>3.0.0-M4</version>
                 <configuration>
                     <dependenciesToScan>
-                        <dependency>jakarta.json:jakarta.json-tck-tests-plugability</dependency>
+                        <dependency>jakarta.json:jakarta.json-tck-tests-pluggability</dependency>
                     </dependenciesToScan>
                     <trimStackTrace>false</trimStackTrace>
                     <failIfNoTests>true</failIfNoTests>
