+++ impl-tck/tck-tests/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2020, 2021 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2020, 2022 Oracle and/or its affiliates. All rights reserved.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -24,7 +24,7 @@
 
     <parent>
         <groupId>org.eclipse.jsonp</groupId>
-        <artifactId>jakarta.json-tck</artifactId>
+        <artifactId>tck-project</artifactId>
         <version>2.1.0-SNAPSHOT</version>
     </parent>
