+++ tck/tck-dist/pom.xml
@@ -21,13 +21,13 @@
    <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>jakarta.json</groupId>
-        <artifactId>jakarta.json-tck</artifactId>
-        <version>2.1.0-SNAPSHOT</version>
+       <artifactId>jakarta.json-tck-project</artifactId>
+       <version>2.1.0-SNAPSHOT</version>
        <relativePath>../pom.xml</relativePath>
     </parent>
 
    <groupId>jakarta.json</groupId>
-   <artifactId>jsonp-tck-dist</artifactId>
+   <artifactId>jakarta.json-tck</artifactId>
    <packaging>pom</packaging>
    <name>JSONP TCK Distribution</name>
 
@@ -40,7 +40,7 @@
       </dependency>
       <dependency>
          <groupId>jakarta.json</groupId>
-         <artifactId>jakarta.json-tck</artifactId>
+         <artifactId>jakarta.json-tck-project</artifactId>
          <version>${project.version}</version>
          <type>pom</type>
       </dependency>
@@ -84,19 +84,19 @@
       </dependency>
       <dependency>
          <groupId>jakarta.json</groupId>
-         <artifactId>jakarta.json-tck-tests-plugability</artifactId>
+         <artifactId>jakarta.json-tck-tests-pluggability</artifactId>
          <version>${project.version}</version>
          <type>jar</type>
       </dependency>
       <dependency>
          <groupId>jakarta.json</groupId>
-         <artifactId>jakarta.json-tck-tests-plugability</artifactId>
+         <artifactId>jakarta.json-tck-tests-pluggability</artifactId>
          <version>${project.version}</version>
          <type>pom</type>
       </dependency>
       <dependency>
          <groupId>jakarta.json</groupId>
-         <artifactId>jakarta.json-tck-tests-plugability</artifactId>
+         <artifactId>jakarta.json-tck-tests-pluggability</artifactId>
          <version>${project.version}</version>
          <classifier>sources</classifier>
          <optional>true</optional>
