+++ tck/tck-dist/src/main/resources/LICENSE_EFTL.md
