+++ tck/tck-docs/userguide/pom.xml
@@ -21,7 +21,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>jakarta.json</groupId>
-        <artifactId>jakarta.json-tck</artifactId>
+        <artifactId>jakarta.json-tck-project</artifactId>
         <version>2.1.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
