+++ tck/tck-tests-pluggability/pom.xml
@@ -21,11 +21,12 @@
 
     <parent>
         <groupId>jakarta.json</groupId>
-        <artifactId>jakarta.json-tck</artifactId>
+        <artifactId>jakarta.json-tck-project</artifactId>
         <version>2.1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>jakarta.json-tck-tests-plugability</artifactId>
+    <artifactId>jakarta.json-tck-tests-pluggability</artifactId>
     <packaging>jar</packaging>
 
     <dependencies>
