+++ tck/tck-tests-pluggability/src/main/java/jakarta/jsonp/tck/provider/MyJsonGeneratorFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2021 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -20,7 +20,6 @@
 
 package jakarta.jsonp.tck.provider;
 
-import jakarta.json.*;
 import jakarta.json.stream.*;
 import java.io.*;
 import java.nio.charset.Charset;
@@ -28,26 +27,26 @@ import java.util.*;
 import java.util.logging.Logger;
 
 /*
- * MyJsonParserFactory is a Json Test ParserFactory used by the pluggability tests
+ * MyJsonGeneratorFactory is a Json Test GeneratorFactory used by the pluggability tests
  * to test the Json SPI layer. This parser tracks that the proper callback
  * methods are invoked within the parser when Json API methods are called.
  */
 
-public class MyJsonParserFactory implements JsonParserFactory {
+public class MyJsonGeneratorFactory implements JsonGeneratorFactory {
 
-  private static final Logger LOGGER = Logger.getLogger(MyJsonParserFactory.class.getName());
+  private static final Logger LOGGER = Logger.getLogger(MyJsonGeneratorFactory.class.getName());
 
-  private InputStream in = null;
+  private OutputStream out = null;
 
-  private Charset charset = null;
+  private Writer writer = null;
 
-  private Reader reader = null;
+  private Charset charset = Charset.forName("UTF-8");
 
   private Map<String, ?> config = null;
 
   private void dumpInstanceVars() {
-    LOGGER.info("reader=" + reader);
-    LOGGER.info("in=" + in);
+    LOGGER.info("writer=" + writer);
+    LOGGER.info("out=" + out);
     LOGGER.info("charset=" + charset);
     LOGGER.info("config=" + config);
   }
@@ -67,7 +66,7 @@ public class MyJsonParserFactory implements JsonParserFactory {
     calls.append(s);
   }
 
-  public MyJsonParserFactory(Map<String, ?> config) {
+  public MyJsonGeneratorFactory(Map<String, ?> config) {
     this.config = config;
   }
 
@@ -77,37 +76,26 @@ public class MyJsonParserFactory implements JsonParserFactory {
     return config;
   }
 
-  public JsonParser createParser(InputStream in) {
-    LOGGER.info("public JsonParser createParser(InputStream)");
-    addCalls("public JsonParser createParser(InputStream)");
-    this.in = in;
+  public JsonGenerator createGenerator(OutputStream out) {
+    LOGGER.info("public JsonGenerator createGenerator(OutputStream)");
+    addCalls("public JsonGenerator createGenerator(OutputStream)");
+    this.out = out;
     return null;
   }
 
-  public JsonParser createParser(InputStream in, Charset charset) {
-    LOGGER.info("public JsonParser createParser(InputStream, Charset)");
-    addCalls("public JsonParser createParser(InputStream, Charset)");
-    this.in = in;
+  public JsonGenerator createGenerator(OutputStream out, Charset charset) {
+    LOGGER.info(
+        "public JsonGenerator createGenerator(OutputStream, Charset)");
+    addCalls("public JsonGenerator createGenerator(OutputStream, Charset)");
+    this.out = out;
     this.charset = charset;
     return null;
   }
 
-  public JsonParser createParser(Reader reader) {
-    LOGGER.info("public JsonParser createParser(Reader)");
-    addCalls("public JsonParser createParser(Reader)");
-    this.reader = reader;
-    return null;
-  }
-
-  public JsonParser createParser(JsonArray jsonArray) {
-    LOGGER.info("public JsonParser createParser(JsonArray)");
-    addCalls("public JsonParser createParser(JsonArray)");
-    return null;
-  }
-
-  public JsonParser createParser(JsonObject jsonObject) {
-    LOGGER.info("public JsonParser createParser(JsonObject)");
-    addCalls("public JsonParser createParser(JsonObject)");
+  public JsonGenerator createGenerator(Writer writer) {
+    LOGGER.info("public JsonGenerator createGenerator(Writer)");
+    addCalls("public JsonGenerator createGenerator(Writer)");
+    this.writer = writer;
     return null;
   }
 }
