+++ tck/tck-tests/pom.xml
@@ -21,8 +21,9 @@
 
     <parent>
         <groupId>jakarta.json</groupId>
-        <artifactId>jakarta.json-tck</artifactId>
+        <artifactId>jakarta.json-tck-project</artifactId>
         <version>2.1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
